// packages/sdk/src/orderUid.ts
export interface CowOrderData {
  sellToken: Address
  buyToken: Address
  receiver: Address
  sellAmount: bigint
  buyAmount: bigint
  validTo: number
  appData: Hex
  feeAmount: bigint
  kind: Hex
  partiallyFillable: boolean
  sellTokenBalance: Hex
  buyTokenBalance: Hex
}